requirements.txt
streamlit
openai
python-dotenv
PyMuPDF 
sentence-transformers 
faiss-cpu
transformers>=4.41.0
datasets
peft>=0.10.0
accelerate
bitsandbytes